Commit:42

+++ app/src/main/java/com/example/vetfinance/data/Product.kt
@@ -15,7 +15,7 @@ data class Product(
     val name: String,
     var price: Double,
     var cost: Double,
-    var stock: Int,
+    var stock: Double,
     val isService: Boolean,
     var sellingMethod: String = SELLING_METHOD_BY_UNIT
 )
